practice/chapter11/code/requirements.txt
# 11장 그래프 분석 실습 의존성
# Python 3.10+

# 그래프 분석
networkx>=3.1
python-louvain>=0.16

# 그래프 임베딩
node2vec>=0.4.6
gensim>=4.3.0

# 추천 시스템
scikit-surprise>=1.1.3

# 데이터 처리 및 시각화
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3

# 평가 지표
scipy>=1.10
